src/test/javascript/spec/app/entities/drink/drink-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import AlertService from '@/shared/alert/alert.service';
import * as config from '@/shared/config/config';
import DrinkUpdateComponent from '@/entities/drink/drink-update.vue';
import DrinkClass from '@/entities/drink/drink-update.component';
import DrinkService from '@/entities/drink/drink.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});

describe('Component Tests', () => {
  describe('Drink Management Update Component', () => {
    let wrapper: Wrapper<DrinkClass>;
    let comp: DrinkClass;
    let drinkServiceStub: SinonStubbedInstance<DrinkService>;

    beforeEach(() => {
      drinkServiceStub = sinon.createStubInstance<DrinkService>(DrinkService);

      wrapper = shallowMount<DrinkClass>(DrinkUpdateComponent, {
        store,
        localVue,
        router,
        provide: {
          alertService: () => new AlertService(store),
          drinkService: () => drinkServiceStub,
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.drink = entity;
        drinkServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(drinkServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.drink = entity;
        drinkServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(drinkServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });
  });
});
